[fix] handling possible segmentation fault
The HWPE spec has some fixed registers for example CLEAR register to
make the hw in a defined state. Some of the registers are not implemented
in the model as commands such as CLEAR is generally done at the beginning
of the execution and in the model it is taken care by the FSM. But it
needs to be handled by a check. Bu default deducting an offset of 0x20
can cause the index of the register to be negative and result in a
segmentation fault